Release/v1.0.2

See merge request we-private/backend/balances-history!7

+++ src/lib/api/config.rs
@@ -16,7 +16,7 @@ struct ConfigFlat {
     pub pgdatabase: String,
     pub pguser: String,
     pub pgpassword: String,
-    pub pgpoolsize: u32,
+    pub pgpoolsize: usize,
     pub pgconnection_timeout: u32,
     pub pgkeepalives_idle: u32,
 }
